backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.2.0
django-cors-headers==3.11.0
gunicorn==20.1.0
django-filter==22.1